requirements.txt
fastapi
uvicorn
passlib
python-dotenv
pymongo
PyJWT
pydantic
make
pydantic[email]